Rename arrhelp to arrutil
